+++ integrations/facebook-pixel/test/index.test.js
@@ -298,7 +298,7 @@ describe('Facebook Pixel', function() {
           assertEventId(window.fbq);
         });
 
-        it('should send whitelisted PII properties', function() {
+        it('should send allowed PII properties', function() {
           facebookPixel.options.whitelistPiiProperties = ['country'];
           analytics.track('event', {
             // PII
@@ -330,7 +330,7 @@ describe('Facebook Pixel', function() {
           assertEventId(window.fbq);
         });
 
-        it('should blacklist properties defined in the blacklistPiiProperties setting', function() {
+        it('should deny properties defined in the blacklistPiiProperties setting', function() {
           facebookPixel.options.blacklistPiiProperties = [
             { propertyName: 'team', hashPropery: false }
           ];
@@ -350,7 +350,7 @@ describe('Facebook Pixel', function() {
             // Non PII
             position: 'point guard',
 
-            // Not default PII but included in blacklist setting
+            // Not default PII but included in denylist setting
             team: 'Warriors'
           });
 
@@ -366,7 +366,7 @@ describe('Facebook Pixel', function() {
           assertEventId(window.fbq);
         });
 
-        it('should hash and send blacklisted properties if the hashProperty flag is true', function() {
+        it('should hash and send denied properties if the hashProperty flag is true', function() {
           facebookPixel.options.blacklistPiiProperties = [
             {
               propertyName: 'email',
@@ -393,7 +393,7 @@ describe('Facebook Pixel', function() {
             // Non PII
             position: 'point guard',
 
-            // Not default PII but included in blacklist setting
+            // Not default PII but included in denylist setting
             team: 'Warriors'
           });